+++ examples/other/trimesh/nearest.py
@@ -26,4 +26,4 @@ cloud_close.vertices_color    = cloud_colors
 arrs = Arrows(cloud_original.vertices, cloud_close.vertices, c='w')
 
 ## create a scene containing the mesh and two sets of points
-show(mesh, cloud_original, cloud_close, arrs, __doc__, bg='bb')
+show(mesh, cloud_original, cloud_close, arrs, __doc__, bg='bb', axes=1, viewup='z')